backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo Backend Application"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlmodel>=0.0.8",
    "asyncpg>=0.28.0",
    "better-auth",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "psycopg2-binary>=2.9.11",
    "openai>=1.12.0",
    "httpx>=0.28.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.24.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
